requirements.txt
line-bot-sdk 
gunicorn
flask
openai